shell/configuration: support unsavedChanges property
plasma-workspace now supports a property unsavedChanges in applet config
pages that can communicate whether the present state is different from
the saved configuration state (in cases where the automatic detection is
not good enough). Plasma-nano always saves automatically, so the feature
doesn't do anything, but it should listen to this property changing to
not miss unsaved changes.

This change makes it save if the unsavedChanges property of a config
page becomes true.